Not relying on Azure Keyvault anymore for signing.
Reason: I have to come up with a backup strategy for the configuration data which is stored in an entity framework DB, so why do it again for the certs stored on azure keyvault.

I create an ECDsa PFX with a password, which is use it or lose it, and store those pfx's in the database.

Also, it looks like signing of 2 jwts takes about 5ms (nice) in the arbitrary_id_token